+++ app/src/main/java/com/agileburo/anytype/ui/page/PageFragment.kt
@@ -15,11 +15,13 @@ import androidx.lifecycle.ViewModelProviders
 import androidx.lifecycle.lifecycleScope
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.agileburo.anytype.R
+import com.agileburo.anytype.core_ui.common.Markup
 import com.agileburo.anytype.core_ui.extensions.invisible
 import com.agileburo.anytype.core_ui.extensions.visible
 import com.agileburo.anytype.core_ui.features.page.BlockAdapter
 import com.agileburo.anytype.core_ui.features.page.BlockView
 import com.agileburo.anytype.core_ui.features.page.styling.StylingEvent
+import com.agileburo.anytype.core_ui.features.page.styling.StylingMode
 import com.agileburo.anytype.core_ui.menu.DocumentPopUpMenu
 import com.agileburo.anytype.core_ui.model.UiBlock
 import com.agileburo.anytype.core_ui.reactive.clicks
@@ -294,13 +296,35 @@ open class PageFragment :
             styleToolbar.events.collect { event ->
                 when (event) {
                     is StylingEvent.Coloring.Text -> {
-                        vm.onMarkupTextColorAction(
-                            color = event.color.title
-                        )
+                        if (styleToolbar.mode == StylingMode.MARKUP)
+                            vm.onMarkupTextColorAction(event.color.title)
+                        else
+                            vm.onToolbarTextColorAction(event.color.title)
                     }
                     is StylingEvent.Coloring.Background -> {
-                        vm.onMarkupBackgroundColorAction(
-                            color = event.color.title
+                        if (styleToolbar.mode == StylingMode.MARKUP)
+                            vm.onMarkupBackgroundColorAction(event.color.title)
+                        else
+                            vm.onBlockBackgroundColorAction(event.color.title)
+                    }
+                    is StylingEvent.Markup.Bold -> {
+                        vm.onBlockStyleMarkupActionClicked(
+                            action = Markup.Type.BOLD
+                        )
+                    }
+                    is StylingEvent.Markup.Italic -> {
+                        vm.onBlockStyleMarkupActionClicked(
+                            action = Markup.Type.ITALIC
+                        )
+                    }
+                    is StylingEvent.Markup.Strikethrough -> {
+                        vm.onBlockStyleMarkupActionClicked(
+                            action = Markup.Type.STRIKETHROUGH
+                        )
+                    }
+                    is StylingEvent.Markup.Code -> {
+                        vm.onBlockStyleMarkupActionClicked(
+                            action = Markup.Type.KEYBOARD
                         )
                     }
                 }